Updated to version 3.08.2 (SVN Revision: 733) for studying the 2D systems of quantum dots.
This version is obtained from the Subversion repository that is currently
used for version control and distribution of CHAMP program (permission is
granted for this commit).

This is not the most recent version, yet it is the most stable, working and
tested version "specifically for 2D systems of quantum dots".

Please contact one of the principal authors, if you would like to use the most
recent version.